scripts/requirements.txt
requests>=2.31.0
feedparser>=6.0.10
openai>=1.30.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
